third_party/WebKit/LayoutTests/fast/peerconnection/RTCPeerConnection-localDescription-expected.txt

Tests RTCPeerConnection localDescription.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testExecutionOrderClosedConnection() did not throw exception.
PASS pc.setLocalDescription().catch(expectedTypeError) did not throw exception.
PASS pc.setLocalDescription(null).catch(expectedInvalidSessionDescription) did not throw exception.
PASS pc.setLocalDescription(sessionDescription, requestSucceeded1, unexpectedCallback); did not throw exception.
PASS error.name is "InvalidStateError"
PASS error.toString() is "InvalidStateError: The RTCPeerConnection's signalingState is 'closed'."
PASS events is [1,2,3]
PASS error.name is "TypeError"
PASS expectedTypeError was called.
PASS error.name is "OperationError"
PASS expectedInvalidSessionDescription was called.
PASS requestSucceeded was called.
PASS pc.setLocalDescription(sessionDescription, unexpectedCallback, requestFailed2); did not throw exception.
PASS requestFailed was called.
PASS pc.localDescription.type is "offer"
PASS pc.localDescription.sdp is "local"
PASS pc.localDescription.type is "offer"
PASS pc.localDescription.sdp is "local"
PASS successfullyParsed is true

TEST COMPLETE
